Revert "Merge branch 'master' of https://github.com/Hellorocio/ProjectHardlight"
This reverts commit 894f851480eddb1b49b611bb05cd376567e18a53, reversing
changes made to a714a7c8ca55b4eb6c7279cb0016d8d1807a555a.

+++ Project Hardlight/Assets/Scripts/Attackable.cs
@@ -21,7 +21,7 @@ public class Attackable : MonoBehaviour
     
     
     public int maxHealth;
-    public int currentHealth;
+    public float currentHealth;
     public Color defaultColor;
 
     //health changed event- used by HealthBar
@@ -43,7 +43,7 @@ public class Attackable : MonoBehaviour
     
     public void Heal (float amt)
     {
-        currentHealth += (int) amt;
+        currentHealth += amt;
         if (currentHealth >= maxHealth)
         {
             currentHealth = maxHealth;
@@ -62,7 +62,7 @@ public class Attackable : MonoBehaviour
         // Calculate based on modifiers
         float realDamage = dmg * (1.0f + percentDamageTakenModifier);
         //Debug.Log(realDamage);
-        currentHealth -= (int) realDamage;
+        currentHealth -= realDamage;
         if (currentHealth < 0)
         {
             currentHealth = 0;
